action-server new changes: --force flag and empty check depends on default exclusion patterns.
